move preferences types to their own file
